CHANGED DEFAULT ORIGSPACE
Before the 2018-05-21, all data from Geoffrey setup were in ALIT (anterior left inferior time) and converted to RAST (right anterior superior time) during analysis.

Since the 2018-05-21, Geoffrey changed it's setup so its output was in ARIT, but we didn't change the space parameter and the program still thought it was ALIT. So when converting to RAS, it was actually converting to LAS. All data are then left-right inverted.

From this commit, the parameter describe well the data.